+++ jest.config.js
@@ -2,8 +2,10 @@ const jestJupyterLab = require('@jupyterlab/testutils/lib/jest-config');
 
 const esModules = [
   '@codemirror',
+  '@jupyter/react-components',
   '@jupyter/ydoc',
   '@jupyterlab/',
+  '@lumino/',
   'lib0',
   'nanoid',
   'vscode-ws-jsonrpc',